requirements.txt
colorgram.py==1.2.0
cycler==0.10.0
ImageHash==4.2.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.2.0
Pillow==8.0.1
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
scipy==1.6.0
six==1.15.0
